upadte go and ts clean
