+++ runner/src/container.ts
@@ -1,0 +1,7 @@
+import { Container } from "@cloudflare/containers";
+
+export class SalamSandbox extends Container {
+  override defaultPort = 8080;
+  override sleepAfter = "2m";
+  override enableInternet = false;
+}
